Add input output notebook
